feat: Переделать выдачу результата с CommentDto на CommentResponseDto
